+++ leviathan/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-__version__ = "2024.7.31"
+__version__ = "2024.9.23"
 from . import utils
 from . import index